+++ tests/Integration/Database/RuleSetTest.php
@@ -38,4 +38,39 @@ final class RuleSetTest extends DatabaseTestCase
 
         $this->assertEquals($ruleSet, self::$db->getRuleSet());
     }
+
+
+    /**
+     * @see https://github.com/kreait/firebase-php/issues/705
+     */
+    public function testRulesAreProperlyEncoded(): void
+    {
+        $rules = RuleSet::private()->getRules();
+        $rules['rules'][self::$refPrefix.__FUNCTION__] = [
+            'value1' => [
+                '.indexOn' => [
+                    'ab'
+                ]
+            ],
+            'value2' => [
+                '.indexOn' => [
+                    'cd',
+                    'ef'
+                ]
+            ],
+        ];
+
+        $ruleSet = RuleSet::fromArray($rules);
+
+        self::$db->updateRules($ruleSet);
+
+        $response = self::$apiClient
+            ->get(
+                self::$db->getReference()->getUri()->withPath('/.settings/rules.json')
+            );
+
+        $this->assertSame(200, $response->getStatusCode());
+        // Assert that the returned JSON doesn't contain objects with integer keys instead of lists
+        $this->assertStringNotMatchesFormat('/\d:/', (string) $response->getBody());
+    }
 }