+++ tests/Integration/DatabaseTestCase.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Tests\Integration;
 
+use GuzzleHttp\Client;
 use Kreait\Firebase\Contract\Database;
 use Kreait\Firebase\Tests\IntegrationTestCase;
 use Kreait\Firebase\Util;
@@ -18,6 +19,8 @@ abstract class DatabaseTestCase extends IntegrationTestCase
 
     protected static Database $db;
 
+    protected static Client $apiClient;
+
     public static function setUpBeforeClass(): void
     {
         parent::setUpBeforeClass();
@@ -29,8 +32,9 @@ abstract class DatabaseTestCase extends IntegrationTestCase
         }
 
         self::$db = self::$factory->withDatabaseUri(self::$rtdbUrl)->createDatabase();
-        self::$refPrefix = 'tests'.\bin2hex(\random_bytes(5));
+        self::$apiClient = self::$factory->createApiClient(['http_errors' => false]);
 
+        self::$refPrefix = 'tests'.\bin2hex(\random_bytes(5));
         self::$db->getReference(self::$refPrefix)->remove();
     }